requirements.txt
flask
requests
pytest
allure-pytest

